[refactor]: 반응형 react-responsive로 변경중
추후 모바일까지 호환을 위해 변경중
현재 홈 , 드로우 페이지 적용
