+++ src/components/assets/draw/Image.jsx
@@ -1,3 +1,6 @@
 import weight from '../draw/weight.svg'
+import hue from '../draw/hue.png';
+import lightness from '../draw/lightness.png'
+import saturation from '../draw/saturation.png';
 
-export{weight}
+export{weight ,hue,lightness,saturation}
