+++ src/components/homePage/Desktop/DeskHome.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import DeskFirstPage from "./DeskFirstPage";
+import DeskSecondPage from "./DeskSecondPage";
+import DeskThirdPage from "./DeskThirdPage";
+import DeskFourthPage from "./DeskFourthPage";
+
+const DeskHome = () => {
+  return (
+    <div>
+      <DeskFirstPage />
+      <DeskSecondPage />
+      <DeskThirdPage />
+      <DeskFourthPage />
+    </div>
+  );
+};
+
+export default DeskHome;
