+++ src/pages/HomePage.jsx
@@ -1,8 +1,43 @@
 import React from "react";
-import Homepage from "../components/homePage/Home";
+import { useMediaQuery } from "react-responsive";
+import DeskHome from "../components/homePage/Desktop/DeskHome";
+import Homepage from "../components/homePage/mobile/Nonhome";
 
-function Home() {
-  return <Homepage />;
-}
+//태블릿 부터 13인치 노트북까지 동일한 UI
+const Desktop = ({ children }) => {
+  const isDesktop = useMediaQuery({ minWidth: 1366 });
+  return isDesktop ? children : null;
+};
+
+const Tablet = ({ children }) => {
+  const isTablet = useMediaQuery({ minWidth: 768, maxWidth: 1366 });
+  return isTablet ? children : null;
+};
+
+const Mobile = ({ children }) => {
+  const isMobile = useMediaQuery({ maxWidth: 767 });
+  return isMobile ? children : null;
+};
+
+const Home = () => {
+  return (
+    <div>
+      <Desktop>
+        <DeskHome />
+      </Desktop>
+
+      <Tablet>
+        <Homepage />
+      </Tablet>
+
+      <Mobile>
+        <Homepage />
+      </Mobile>
+
+      {/*
+      <Default>Not mobile (desktop or laptop or tablet)</Default> */}
+    </div>
+  );
+};
 
 export default Home;
